software/build: patch cmake versions
CMake deprecated a few options that were completely removed in the
latest version 4. In their infinite wisdom, they decided to treat the
"minimum CMake version" as "this project definitely uses features that
old and that have been removed". They decided to treat a baseline
setting as a timebomb. And it's now a hard error to require a version
older than 3.5 as a *baseline*.

As it happens, Arch Linux recently upgraded to CMake 4 (4.0.2-dirty to
be specific, which I find highly ironic).

lzo and read-edid are such projects that can build with old CMake
version. Just patch them as they work just fine under version 4.