software: reduce kernel memory footprint to boot on F1C100s
By upgrading from 5.11 to 6.14, the kernel failed to boot. This is due
to a few things:

- the kernel grew from 5.11 to 6.14
- it now eagerly resolves the device tree early on boot, triggering out
  of memory conditions

This can easily be fixed by disabling CMA (Contiguous Memory Allocation)
that needed for some DMA types. This is not needed here, and even if it
were needed, we don't need 16MB of it. So straight up turn it off.

Also, bump to kernel 6.15, just because.