software: bump up the F1C100s video resolution
By disabling CMA, we free up enough memory that we can use 854x480
resolution of F1C100s, which happens to be a standard resolution
(FWVGA). This helps monitors and the SiI9022 work properly.

We could bump up to 960x540, but this keeps DOOM at x2 scaling and has
too much borders around to be worth it.